requirements.txt
pyyaml==6.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-ecs==4.24.19